fix : change column database

+++ database/migrations/2024_05_15_210821_change_no_antrian_to_pendaftaran_pasien_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('pendaftaran_pasien', function (Blueprint $table) {
-            $table->bigInteger('no_antrian')->nullable()->change();
+            $table->smallInteger('no_antrian')->nullable()->change();
         });
     }
