+++ database/migrations/2024_05_15_211706_change_kuota_to_dokter_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('dokter', function (Blueprint $table) {
-            $table->bigInteger('kuota')->nullable()->default(null)->change();
+            $table->smallInteger('kuota')->nullable()->default(null)->change();
         });
     }
