+++ database/migrations/2024_05_27_141712_add_kuota_bpjs_to_dokter_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('dokter', function (Blueprint $table) {
-            $table->bigInteger('kuota_bpjs')->nullable()->after('kuota');
+            $table->smallInteger('kuota_bpjs')->nullable()->after('kuota');
         });
     }
